refactor for map background
