+++ genetic-algorithms/src/app/app.component.html
@@ -4,3 +4,4 @@
 
 
 
+
