+++ genetic-algorithms/src/libs/features/homegrid/homegrid.component.html
@@ -1,41 +1,7 @@
-<div class="wrapper">
-    <header class="header">
-        <div class="searchDiv">
-            <app-search-bar 
-                [latitude]="controlService.latitude" 
-                [longitude]="controlService.longitude"
-                [cityName]="controlService.cityName" >
-            </app-search-bar>
-        </div>
+<div class="mapBackground">
+    <app-map></app-map>
+</div>
 
-    </header>
-
-    <div class="left">
-        <div class="listDiv">
-            <app-locations-list></app-locations-list>
-        </div>
-
-    </div>
-    
-    <div class="center">
-        <div class="agm-map">
-            <app-map 
-                [searchString]="controlService.searchString">
-            </app-map>
-        </div>
-    </div>
-    
-    <div class="right">
-        <div class="outputDiv">
-            <app-output-console></app-output-console>
-        </div>
-
-    </div>
-    
-    <div class="footer">
-        <div class="timelineDiv">
-            <app-timeline></app-timeline>
-        </div>
-        
-    </div>
+<div class="searchBar">
+    <app-search-bar></app-search-bar>
 </div>
