+++ genetic-algorithms/src/libs/features/homegrid/homegrid.component.scss
@@ -1,56 +1,17 @@
-.wrapper {
-    display: grid;
-    grid-template: 1fr 2fr 1fr / 1fr 2fr 1fr;
-    max-height: calc(100vh - 10%);
-    min-height: calc(100vh - 10%);
-}
-
-.header {
-    border: 2px solid pink;
-    grid-column: 2 / 3;
-    padding: 20px;
-    display: grid;
-    place-items: center;
-}
-
-.left {
-    border: 2px solid blue;
-    grid-row: 1 / 3;
-    grid-column: 1 / 2;
-    padding: 20px;
-    display: grid;
-    justify-content: center;
-}
-
-
-.center {
-    border: 2px solid red;
-    grid-column: 2 / 3;
-    padding: 10px;
-    display: grid;
-    place-items: center;
-
-}
-
 .agm-map {
-    padding: 10px;
-    width: inherit;
     aspect-ratio: 16 / 9;
 }
 
-
-.right {
-    border: 2px solid orange;
-    grid-row: 1 / 3;
-    grid-column: 3 / 4;
-    padding: 20px;
-
+.searchBar {
+    position:absolute;
+    margin-top: 7.5%;
+    z-index: 100;
+    border: 1px solid black;
+    padding: 10px;
+    width: 60%;
+    left: 20%;
+    background-color: lightgray;
+    border-radius: 10px;
 }
 
-.footer {
-    border: 2px solid green;
-    grid-column: 1 / 4;
-    padding: 20px;
-    display: grid;
-    place-items: center;
-}
+
