+++ genetic-algorithms/src/libs/features/map/map.component.html
@@ -11,15 +11,10 @@
         [longitude]="controlService.longitude">
     </agm-marker>
     <agm-polyline  [editable]="false">
-        <agm-polyline-point *ngFor="let data of lines"
+        <agm-polyline-point *ngFor="let data of controlService.lines"
         [latitude]="data.latitude"
         [longitude]="data.longitude">
         </agm-polyline-point>
     </agm-polyline>
     </agm-map>
   </div>
-
-  <div class="optimize" >
-      <button (click)="getOptimizedPath()">Optimize my path!</button>
-  </div>
-
