+++ genetic-algorithms/src/libs/features/map/map.component.ts
@@ -1,12 +1,6 @@
-import { Component, NgZone, OnInit, ViewChild, ElementRef, Input } from '@angular/core';
-import { DistanceParsingService } from '../../services/distance-parsing.service';
-import { DistanceMatrix, mockLocationsList } from '../../../assets/testFolder/testLocations';
-import { litterHistory5 } from '../../../assets/testFolder/litter-history';
+import { Component, OnInit, Input } from '@angular/core';
 import { MapsAPILoader } from '@agm/core';
-import { HttpClient, HttpHeaders, HttpParams } from '@angular/common/http'
-import { apiKey } from 'src/libs/services/secrets/secret.service';
 import { ControlService } from 'src/libs/services/control.service';
-import { switchMapTo, switchMap, take } from 'rxjs/operators'
 
 @Component({
   selector: 'app-map',
@@ -34,74 +28,11 @@ export class MapComponent implements OnInit {
     })
   }
 
-  constructor( 
-   private distanceParser: DistanceParsingService,
+  constructor(
    private mapsAPILoader: MapsAPILoader,
-   private ngZone: NgZone,
-   private httpService: HttpClient,
    public controlService: ControlService
   ) {}
 
-  addLocations(route) {
-    this.lines = []
-    for (let index of route) {
-      this.lines.push(this.controlService.locationsList[index])
-    }
-    this.lines.push(this.controlService.locationsList[route[0]]);
-  }
-
-  async showHistory(litterHistory) {
-    for (let i in litterHistory) {
-      await this.delay(100)
-      this.addLocations(litterHistory[i][0])
-      console.log('Gen: ', i)
-      console.log('Distance: ', litterHistory[i][1])
-    }
-  }
-
-  delay(ms: number) {
-    return new Promise(resolve => setTimeout(resolve, ms));
-  }
-
-  getHttpRequestBody(): string {
-    let origins = this.controlService.locationsList.map(place => place.name);
-    const reqBody = origins.join(' | ');
-    console.log(reqBody);
-    return reqBody
-  }
-
-  callDistanceMatrix$() {
-    const placesParam = this.getHttpRequestBody();
-    const headers = new HttpHeaders().set('Access-Control-Allow-Origin', '*')
-    const params = new HttpParams().set('origins', placesParam).set('destinations', placesParam).set('key', apiKey)
-    console.log('request: ', params)
-    return this.httpService.get<DistanceMatrix>('/maps/api/distancematrix/json', {
-      headers: headers,
-      params: params
-    })
-  }
-
-  getDistanceMatrix() {
-    return this.callDistanceMatrix$().pipe(take(1)).subscribe((matrix) => console.log(this.distanceParser.parse(matrix)), error => console.error(error))
-  }
-
-  callOptimizePath$(distance_dict) {
-    const headers = new HttpHeaders().set('Content-Type', 'application/json')
-    const params = new HttpParams().set('distance_dict', JSON.stringify(distance_dict))
-    return this.httpService.post<any[]>('/optimize-path', {
-      headers: headers, 
-      params: params
-    })
-  }
-
-  getOptimizedPath() {
-    this.callDistanceMatrix$().pipe(switchMap((distance_dict) => 
-      this.callOptimizePath$(this.distanceParser.parse(distance_dict)))).subscribe((litter) => {
-        console.log(litter);
-        console.log('Best path', litter[-1]);
-        this.showHistory(litter);
-      })
-  }
 
   // Get Current Location Coordinates
   private setCurrentLocation() {
