+++ genetic-algorithms/src/libs/features/search-bar/search-bar.component.html
@@ -1,9 +1,11 @@
 <div class="form-group">
-    <div class="form-group">
-        <label>Enter address</label>
-        <input type="text" class="form-control" (keydown.enter)="$event.preventDefault()" placeholder="Search Nearest Location" autocorrect="off" autocapitalize="off" spellcheck="off" type="text" #search>
-      </div>
-    <button class="submitBtn" (click)="addToList()">
-        Add Location
-    </button>
+    <label>Enter address:</label>
+    <input type="text" class="form-control"
+     (keydown.enter)="$event.preventDefault()" 
+     placeholder="Search Nearest Location" 
+     autocorrect="off" autocapitalize="off" 
+     spellcheck="off" type="text" #search>
+    <button mat-raised-button color="secondary" class="submitBtn" (click)="addToList()">Add Location</button>
 </div>
+    
+
