+++ genetic-algorithms/src/styles.scss
